requirements.txt
torch>=2.3,<2.6
torchvision>=0.18,<0.20
tensorboard>=2.16,<2.21

numpy>=1.24
pillow>=10.0
pyyaml>=6.0
tqdm>=4.66
matplotlib>=3.8